fix warnings, ceil sqrts

+++ euler579c/massiveinteger.cpp
@@ -56,7 +56,7 @@ massiveinteger operator*(const massiveinteger & m1, const massiveinteger & m2)
 	return total;
 }
 
-void massiveinteger::truncate(long n)
+void massiveinteger::truncate(size_t n)
 {
 	if (digits.size() >= n)
 	{
